refactor: isolate FormGroupContext file
