+++ src/FormGroup/FormGroup.tsx
@@ -2,6 +2,7 @@ import React, { useMemo } from 'react';
 import Box, { BoxProps } from '@/internals/Box';
 import { forwardRef } from '@/internals/utils';
 import { useStyles, useCustom, useUniqueId } from '@/internals/hooks';
+import { FormGroupContext } from './FormGroupContext';
 
 export interface FormGroupProps extends BoxProps {
   /**
@@ -10,39 +11,6 @@ export interface FormGroupProps extends BoxProps {
    */
   controlId?: string;
 }
-
-const FormGroupContext = React.createContext<{ controlId?: string }>({});
-
-export const useFormGroup = (controlId?: string) => {
-  const context = React.useContext(FormGroupContext);
-  const fallbackId = useUniqueId('rs-');
-
-  const id = controlId || context.controlId || fallbackId;
-  const helpTextId = `${id}-help-text`;
-  const labelId = `${id}-label`;
-  const errorMessageId = `${id}-error-message`;
-
-  return {
-    /**
-     * The `id` of the `<Form.Control>` component.
-     */
-    controlId: id,
-    /**
-     * The `id` of the `<Form.HelpText>` component.
-     */
-    helpTextId,
-
-    /**
-     * The `id` of the `<Form.ControlLabel>` component.
-     */
-    labelId,
-
-    /**
-     * The `id` of the `<Form.ErrorMessage>` component.
-     */
-    errorMessageId
-  };
-};
 
 /**
  * The `<Form.Group>` component is the easiest way to add some structure to forms.
