+++ src/FormGroup/FormGroupContext.tsx
@@ -1,0 +1,3 @@
+import React from 'react';
+
+export const FormGroupContext = React.createContext<{ controlId?: string }>({});
